[WIP] onUpdate, onDelete & onCreate tests in .ts
