+++ functions/.runtimeconfig.json
@@ -1,0 +1,13 @@
+{
+  "env": {
+    "testingConfig": {
+      "projectId": "react-firebase-admin-eeac2",
+      "serviceAccountKey": "../serviceAccountKey.json",
+      "storageBucket": "react-firebase-admin-eeac2.appspot.com",
+      "databaseURL": "https://react-firebase-admin-eeac2.firebaseio.com"
+    }
+  },
+  "signinlink": {
+    "url": "https://react-firebase-admin-eeac2.firebaseapp.com/login"
+  }
+}
