+++ functions/env.json
@@ -1,0 +1,8 @@
+{
+  "testingConfig": {
+    "databaseURL": "https://react-firebase-admin-eeac2.firebaseio.com",
+    "storageBucket": "react-firebase-admin-eeac2.appspot.com",
+    "projectId": "react-firebase-admin-eeac2",
+    "serviceAccountKey": "../serviceAccountKey.json"
+  }
+}
