+++ functions/package.json
@@ -9,7 +9,7 @@
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log",
     "setup-firebase": "node setupProject.js",
-    "test": "mocha --reporter spec  --timeout 10000"
+    "test": "mocha -r ts-node/register test/**/*.test.ts --timeout 10000"
   },
   "engines": {
     "node": "10"
@@ -26,7 +26,13 @@
     "camelcase": "^5.3.1"
   },
   "devDependencies": {
+    "@types/chai": "^4.2.11",
+    "@types/chai-as-promised": "^7.1.2",
+    "@types/mocha": "^7.0.2",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
     "mocha": "^7.2.0",
+    "ts-node": "^8.10.2",
     "tslint": "^5.12.0",
     "typescript": "^3.8.0"
   },
